+++ tests/snapshots/snap_test_gdpr_api_query.py
@@ -6,6 +6,22 @@ from snapshottest import Snapshot
 
 snapshots = Snapshot()
 
+snapshots["test_gdpr_url_pattern_can_be_configured 1"] = {
+    "children": [
+        {"key": "MEMO", "value": "Memo"},
+        {
+            "children": [
+                {"key": "FIRST_NAME", "value": "First name"},
+                {"key": "LAST_NAME", "value": "Last name"},
+            ],
+            "key": "USER",
+        },
+        {"key": "USER", "value": "First name Last name"},
+        {"children": [], "key": "EXTRA_DATA"},
+    ],
+    "key": "PROFILE",
+}
+
 snapshots["test_get_profile_information_from_gdpr_api 1"] = {
     "children": [
         {"key": "MEMO", "value": "Memo"},