+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `GDPR_API_URL_PATTERN` setting for configuring the GDPR API URL pattern.
 - `GDPR_API_MODEL_LOOKUP` setting for configuring how the GDPR model instance is found.
 - `GDPR_API_USER_PROVIDER` setting for configuring how to obtain a `User` instance from the GDPR model instance.
 - `GDPR_API_DELETER` setting for configuring how data deletion is performed.